+++ src/main/java/com/corundumstudio/socketio/scheduler/CancelableScheduler.java
@@ -16,21 +16,57 @@
 package com.corundumstudio.socketio.scheduler;
 
 import io.netty.channel.ChannelHandlerContext;
+import io.netty.util.HashedWheelTimer;
+import io.netty.util.Timeout;
+import io.netty.util.TimerTask;
+import io.netty.util.internal.PlatformDependent;
 
+import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
 
-public interface CancelableScheduler {
+//Change Interface To Abstract Class
+public abstract class CancelableScheduler {
 
-    void update(ChannelHandlerContext ctx);
+    protected final ConcurrentMap<SchedulerKey, Timeout> scheduledFutures = PlatformDependent.newConcurrentHashMap();
+    protected final HashedWheelTimer executorService;
+	
+    protected volatile ChannelHandlerContext ctx;
 
-    void cancel(SchedulerKey key);
+    public CancelableScheduler() {
+        executorService = new HashedWheelTimer();
+    }
+    
+    public CancelableScheduler(ThreadFactory threadFactory) {
+        executorService = new HashedWheelTimer(threadFactory);
+    }    
+    
+    public void update(ChannelHandlerContext ctx) {
+        this.ctx = ctx;
+    }
+    
+    public void cancel(SchedulerKey key) {
+    	final Timeout timeout = scheduledFutures.remove(key);
+        if (timeout != null) {
+            timeout.cancel();
+        }
+    }
+    
+    public void schedule(final Runnable runnable, long delay, TimeUnit unit) {
+        executorService.newTimeout(new TimerTask() {
+            @Override
+            public void run(Timeout timeout) throws Exception {
+                runnable.run();
+            }
+        }, delay, unit);
+    }
+    
+    abstract void schedule(SchedulerKey key, Runnable runnable, long delay, TimeUnit unit);
 
-    void scheduleCallback(SchedulerKey key, Runnable runnable, long delay, TimeUnit unit);
+    abstract void scheduleCallback(SchedulerKey key, Runnable runnable, long delay, TimeUnit unit);
 
-    void schedule(Runnable runnable, long delay, TimeUnit unit);
-
-    void schedule(SchedulerKey key, Runnable runnable, long delay, TimeUnit unit);
-
-    void shutdown();
-
+    public void shutdown() {
+        executorService.stop();
+    }
+    
 }
