+++ src/main/java/com/corundumstudio/socketio/scheduler/HashedWheelScheduler.java
@@ -15,52 +15,23 @@
  */
 package com.corundumstudio.socketio.scheduler;
 
-import java.util.Map;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
 
 import io.netty.channel.ChannelHandlerContext;
-import io.netty.util.HashedWheelTimer;
 import io.netty.util.Timeout;
 import io.netty.util.TimerTask;
-import io.netty.util.internal.PlatformDependent;
 
-public class HashedWheelScheduler implements CancelableScheduler {
-
-    private final Map<SchedulerKey, Timeout> scheduledFutures = PlatformDependent.newConcurrentHashMap();
-    private final HashedWheelTimer executorService;
-    
-    public HashedWheelScheduler() {
-        executorService = new HashedWheelTimer();
-    }
-    
-    public HashedWheelScheduler(ThreadFactory threadFactory) {
-        executorService = new HashedWheelTimer(threadFactory);
-    }
+public class HashedWheelScheduler extends CancelableScheduler {
 
     private volatile ChannelHandlerContext ctx;
 
-    @Override
-    public void update(ChannelHandlerContext ctx) {
-        this.ctx = ctx;
+    public HashedWheelScheduler() {
+        super();
     }
-
-    @Override
-    public void cancel(SchedulerKey key) {
-        Timeout timeout = scheduledFutures.remove(key);
-        if (timeout != null) {
-            timeout.cancel();
-        }
-    }
-
-    @Override
-    public void schedule(final Runnable runnable, long delay, TimeUnit unit) {
-        executorService.newTimeout(new TimerTask() {
-            @Override
-            public void run(Timeout timeout) throws Exception {
-                runnable.run();
-            }
-        }, delay, unit);
+    
+    public HashedWheelScheduler(ThreadFactory threadFactory) {
+        super(threadFactory);
     }
 
     @Override
@@ -81,9 +52,7 @@ public class HashedWheelScheduler implements CancelableScheduler {
             }
         }, delay, unit);
 
-        if (!timeout.isExpired()) {
-            scheduledFutures.put(key, timeout);
-        }
+        addScheduledFuture(key, timeout);
     }
 
     @Override
@@ -99,9 +68,7 @@ public class HashedWheelScheduler implements CancelableScheduler {
             }
         }, delay, unit);
 
-        if (!timeout.isExpired()) {
-            scheduledFutures.put(key, timeout);
-        }
+        addScheduledFuture(key, timeout);
     }
 
     @Override
@@ -109,4 +76,10 @@ public class HashedWheelScheduler implements CancelableScheduler {
         executorService.stop();
     }
 
+    private void addScheduledFuture(final SchedulerKey key, Timeout timeout){
+        if (!timeout.isExpired()) {
+            scheduledFutures.put(key, timeout);
+        }
+    }
+    
 }
