+++ src/main/java/com/corundumstudio/socketio/scheduler/HashedWheelTimeoutScheduler.java
@@ -25,51 +25,22 @@
 package com.corundumstudio.socketio.scheduler;
 
 import io.netty.channel.ChannelHandlerContext;
-import io.netty.util.HashedWheelTimer;
 import io.netty.util.Timeout;
 import io.netty.util.TimerTask;
-import io.netty.util.internal.PlatformDependent;
 
-import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.TimeUnit;
 
-public class HashedWheelTimeoutScheduler implements CancelableScheduler {
-
-    private final ConcurrentMap<SchedulerKey, Timeout> scheduledFutures = PlatformDependent.newConcurrentHashMap();
-    private final HashedWheelTimer executorService;
+public class HashedWheelTimeoutScheduler extends CancelableScheduler {
 
     private volatile ChannelHandlerContext ctx;
     
     public HashedWheelTimeoutScheduler() {
-        executorService = new HashedWheelTimer();
+       super();
     }
     
     public HashedWheelTimeoutScheduler(ThreadFactory threadFactory) {
-        executorService = new HashedWheelTimer(threadFactory);
-    }
-
-    @Override
-    public void update(ChannelHandlerContext ctx) {
-        this.ctx = ctx;
-    }
-
-    @Override
-    public void cancel(SchedulerKey key) {
-        Timeout timeout = scheduledFutures.remove(key);
-        if (timeout != null) {
-            timeout.cancel();
-        }
-    }
-
-    @Override
-    public void schedule(final Runnable runnable, long delay, TimeUnit unit) {
-        executorService.newTimeout(new TimerTask() {
-            @Override
-            public void run(Timeout timeout) throws Exception {
-                runnable.run();
-            }
-        }, delay, unit);
+        super(threadFactory);
     }
 
     @Override
@@ -109,10 +80,6 @@ public class HashedWheelTimeoutScheduler implements CancelableScheduler {
         replaceScheduledFuture(key, timeout);
     }
 
-    @Override
-    public void shutdown() {
-        executorService.stop();
-    }
 
     private void replaceScheduledFuture(final SchedulerKey key, final Timeout newTimeout) {
         final Timeout oldTimeout;
